temp/requirement.txt
opencv-python
scipy
matplotlib
tqdm
pyyaml